move logger back
